feat(ui): AI 页两端渐进模糊 + 文件管理器列表视图 + 页头安全区修复
- AI 页：消息区成为唯一滚动容器，页头与底栏（含输入栏）粘在其内部，
  内容从两端的渐进模糊里滚过；模糊抽成可复用的 ProgressiveBlur 组件
- PageHeader：粘滞偏移改为 var(--ph-stick)（默认 --safe-top），修掉
  「向上滑动后标题压进手机顶部消息栏」；遮罩向上铺设量随之参数化
- 文件管理器：顶层目录从方块网格改为一行一个的目录列表（图标 / 说明 / 条目数），
  子目录增加「..」上一级行与底部「N 个目录 · M 个文件 · 路径」状态行；
  空目录也保留回退行
- 验证：无头 Edge 结构断言 16/16（遮罩层数、粘滞位置、内容穿行、空目录回退）

注：本提交为工作区快照，同时带上工作区里其他未提交的改动
（AI 提示词的输入格式约定、若干 e2e / 预览脚本、Rust 侧小修）。

+++ src-tauri/src/modules/kb/search.rs
@@ -493,7 +493,7 @@ pub fn search(
 
 /// 无查询词的「浏览」模式：按日期倒序列出某类内容。AI 问「最近有什么」时用。
 pub fn browse(conn: &Connection, q: &KbQuery, enabled: &[String]) -> Result<Vec<KbHit>> {
-    let limit = q.limit.unwrap_or(8).clamp(1, 30) as i64;
+    let limit = q.limit.unwrap_or(8).clamp(1, 30);
     let filter = build_filter(q, enabled);
     let sql = format!(
         "SELECT id, source_type, source_id, title, summary, occurred_on, tags,
